add SendOptions class
